REFACTOR test classes merged from spatie/laravel-honeypot to 'tests/Spatie' directory
