+++ tests/Spatie/HoneypotBladeComponentTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Sfneal\Honeypot\Tests;
+namespace Sfneal\Honeypot\Tests\Spatie;
 
 use Illuminate\Foundation\Testing\Concerns\InteractsWithViews;
+use Sfneal\Honeypot\Tests\TestCase;
 use Spatie\Snapshots\MatchesSnapshots;
 use Spatie\TestTime\TestTime;
