+++ tests/Spatie/HoneypotBladeDirectiveTest.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Sfneal\Honeypot\Tests;
+namespace Sfneal\Honeypot\Tests\Spatie;
 
 use Carbon\CarbonImmutable;
 use Illuminate\Support\DateFactory;
+use Sfneal\Honeypot\Tests\TestCase;
 use Spatie\Snapshots\MatchesSnapshots;
 use Spatie\TestTime\TestTime;
